requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
attrs==23.2.0
binaryornot==0.4.4
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.3
flet==0.22.1
flet-core==0.22.1
flet-runtime==0.22.1
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
orjson==3.10.3
outcome==1.3.0.post0
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
pyparsing==3.1.2
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
PyYAML==6.0.1
qrcode==7.4.2
repath==0.9.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
selenium==4.19.0
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
text-unidecode==1.3
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
watchdog==4.0.0
watchfiles==0.21.0
webdriver-manager==4.0.1
websockets==12.0
wsproto==1.2.0
